chore: update theme config
