+++ src/components/App/index.tsx
@@ -1,16 +1,19 @@
-import Button from '@mui/material/Button';
-import React from 'react';
-import EditIcon from '@mui/icons-material/Edit';
-import CloseIcon from '@mui/icons-material/Close';
 import AddIcon from '@mui/icons-material/Add';
+import Button from '@mui/material/Button';
+import CloseIcon from '@mui/icons-material/Close';
 import DeleteIcon from '@mui/icons-material/Delete';
+import EditIcon from '@mui/icons-material/Edit';
 import SearchIcon from '@mui/icons-material/Search';
 
 const App = () => {
   return (
     <>
       <h1>Start Project</h1>
-      <Button variant="contained" color="secondary">
+      <Button
+        variant="contained"
+        color="secondary"
+        sx={{ fontSize: 'typography.h1' }}
+      >
         Create account{' '}
       </Button>
       <Button variant="outlined" color="secondary">
