+++ src/theme/breakpoints/index.ts
@@ -1,0 +1,9 @@
+import createBreakpoints from '@mui/system/createTheme/createBreakpoints';
+
+export const breakpoints = createBreakpoints({
+  values: {
+    mobile: 320,
+    tablet: 768,
+    desktop: 1280,
+  },
+});
