+++ src/theme/typography/constants.ts
@@ -1,0 +1,11 @@
+export enum FontWeights {
+  THIN = 100,
+  EXTRALIGHT = 200,
+  LIGHT = 300,
+  REGULAR = 400,
+  MEDIUM = 500,
+  SEMIBOLD = 600,
+  BOLD = 700,
+  EXTRABOLD = 800,
+  BLACK = 900,
+}
